fix bilee service payments

+++ src/api/schema/payment.py
@@ -8,5 +8,5 @@ class TopUpMethod(Enum):
     CARD = 'card'
 
 class TopUpSchema(BaseModel):
-    amount: float = Field(ge=10, le=50000)
+    amount: int = Field(ge=10, le=50000)
     method: TopUpMethod = Field(default=TopUpMethod.CARD)
